Supressed css warnings. Fixed markdown display in PSS, breadcrumbs, etc. Fixed module import issue for external links functions.
