+++ components/HomePageComponents/HomePageSlider/index.js
@@ -71,7 +71,7 @@ const HomePageSlider = ({
                   <ReactMarkdown
                     className={css.itemText}
                     source={name}
-                    allowedTypes={["emphasis"]}
+                    allowedTypes={["emphasis", "text"]}
                     unwrapDisallowed
                   />
                 </a>
