+++ lib/index.js
@@ -7,7 +7,8 @@ import decodeHTMLEntities from "./decodeHTMLEntities";
 import deepCopyObject from "./deepCopyObject";
 import endsWith from "./endsWith";
 import escapeForRegex from "./escapeForRegex";
-import externalLinks from "./externalLinks";
+import markdownLinks from "./markdownLinks";
+import wordpressLinks from "./wordpressLinks";
 import extractItemId from "./extractItemId";
 import extractSourceId from "./extractSourceId";
 import extractSourceSetSlug from "./extractSourceSetSlug";
@@ -45,7 +46,8 @@ export {
   deepCopyObject,
   endsWith,
   escapeForRegex,
-  externalLinks,
+  markdownLinks,
+  wordpressLinks,
   extractItemId,
   extractSourceId,
   extractSourceSetSlug,
