+++ lib/markdownLinks.js
@@ -9,23 +9,6 @@ const markdownLinks = reference => {
   );
 };
 
-export const wordpressLinks = links => {
-  for (var i = 0; i < links.length; i++) {
-    var link = links.item(i);
-    if (
-      !link.classList.contains("btn") &&
-      link.href !== "" &&
-      link.hostname !== "dp.la" &&
-      link.hostname !== "pro.dp.la" &&
-      link.hostname !== "pro-staging.dp.la" &&
-      link.hostname !== "ci-test.dp.la" &&
-      link.hostname !== "staging.dp.la" &&
-      link.hostname !== "localhost"
-    )
-      link.classList.add("external");
-  }
-};
-
 const isExternal = href =>
   href.indexOf("/") !== 0 &&
   href.indexOf("http") === 0 &&
