+++ package.json
@@ -80,6 +80,7 @@
     "sinon": "^7.3.2",
     "url-loader": "^2.1.0",
     "webpack": "^4.38.0",
+    "webpack-filter-warnings-plugin": "^1.2.1",
     "webpack-manifest-plugin": "^2.0.4"
   },
   "eslintConfig": {
