+++ pages/about/index.js
@@ -18,11 +18,10 @@ import {
 import {
   getBreadcrumbs,
   getItemWithId,
-  getItemWithName,
+  wordpressLinks,
   getMenuItemUrl,
   decodeHTMLEntities
 } from "lib";
-import { wordpressLinks } from "lib/externalLinks";
 
 import utils from "stylesheets/utils.scss";
 import contentCss from "stylesheets/content-pages.scss";
