+++ pages/primary-source-sets/set/additional-resources.js
@@ -9,8 +9,7 @@ import SourceSetInfo from "components/PrimarySourceSetsComponents/SingleSet/Sour
 import ResourcesTabs from "components/PrimarySourceSetsComponents/SingleSet/ResourcesTabs";
 
 import { PSS_BASE_URL } from "constants/env";
-import { getCurrentFullUrl, removeQueryParams } from "lib";
-import { markdownLinks } from "lib/externalLinks";
+import { getCurrentFullUrl, removeQueryParams, markdownLinks } from "lib";
 
 import utils from "stylesheets/utils.scss";
 import contentCss from "stylesheets/content-pages.scss";
